feat: MCP23017의 GPIOA,B를 모두 초기화 할 수 있게 명령어 호출 추가
GPIOA,B의 핀 사용 용도를 한 변수로 관리하기 위해 pin_status 멤버를
 uint16_t 변수로 변경함

+++ App/drivers/mcp23017.h
@@ -22,7 +22,7 @@ typedef enum {
 typedef struct {
     const uint8_t gpio_pin;
     const uint8_t address;
-    const uint8_t pin_status;
+    const uint16_t pin_status;
 } Mcp23017DeviceConfig;
 
 typedef struct {
